Push out AsPgCStr and PgNode
